requirements.txt
django>=1.8.0
# Additional requirements go here

djangorestframework
